+++ tests/regression/test_static_resolve_all.py
@@ -48,7 +48,7 @@ def test_static_resolve_all():
     with pytest.raises(UnsolvableReturnTypeError):
         dg.node(update_user)
 
-    dg.static_resolve_all()
+    dg.analyze_nodes()
 
 
 def test_dg_node_builtin():